+++ nflow-engine/src/main/java/io/nflow/engine/config/db/OracleDatabaseConfiguration.java
@@ -171,5 +171,15 @@ public class OracleDatabaseConfiguration extends DatabaseConfiguration {
     public boolean useBatchUpdate() {
       return useBatchUpdate;
     }
+
+    @Override
+    public String clobToComparable(String column) {
+      return "to_char(" + column + ")";
+    }
+
+    @Override
+    public String forUpdateSkipLocked() {
+      return "";
+    }
   }
 }